Move service definitions into their respective modules
