+++ custom_components/ramses_cc/climate.py
@@ -1,7 +1,7 @@
 """Support for RAMSES climate entities."""
 from __future__ import annotations
 
-from datetime import datetime
+from datetime import datetime, timedelta
 import json
 import logging
 from typing import Any
@@ -9,6 +9,7 @@ from typing import Any
 from ramses_rf.system.heat import Evohome
 from ramses_rf.system.zones import Zone
 from ramses_tx.const import SZ_MODE, SZ_SETPOINT, SZ_SYSTEM_MODE
+import voluptuous as vol
 
 from homeassistant.components.climate import (
     DOMAIN as PLATFORM,
@@ -29,22 +30,42 @@ from homeassistant.components.climate import (
 )
 from homeassistant.const import UnitOfTemperature
 from homeassistant.core import HomeAssistant, callback
-from homeassistant.helpers import entity_platform
+from homeassistant.helpers import config_validation as cv, entity_platform
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
 from . import RamsesEntity
 from .const import (
+    ATTR_DURATION,
+    ATTR_LOCAL_OVERRIDE,
+    ATTR_MAX_TEMP,
+    ATTR_MIN_TEMP,
+    ATTR_MODE,
+    ATTR_MULTIROOM,
+    ATTR_OPENWINDOW,
+    ATTR_PERIOD,
+    ATTR_SCHEDULE,
+    ATTR_SETPOINT,
+    ATTR_TEMPERATURE,
+    ATTR_UNTIL,
     BROKER,
     DOMAIN,
     PRESET_CUSTOM,
     PRESET_PERMANENT,
     PRESET_TEMPORARY,
+    SVC_GET_ZONE_SCHEDULE,
+    SVC_PUT_ZONE_TEMP,
+    SVC_RESET_SYSTEM_MODE,
+    SVC_RESET_ZONE_CONFIG,
+    SVC_RESET_ZONE_MODE,
+    SVC_SET_SYSTEM_MODE,
+    SVC_SET_ZONE_CONFIG,
+    SVC_SET_ZONE_MODE,
+    SVC_SET_ZONE_SCHEDULE,
     SystemMode,
     ZoneMode,
 )
 from .coordinator import RamsesBroker
-from .schemas import SVCS_CLIMATE_EVO_TCS, SVCS_CLIMATE_EVO_ZONE
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -95,6 +116,108 @@ PRESET_ZONE_TO_HA = {
 }
 PRESET_HA_TO_ZONE = {v: k for k, v in PRESET_ZONE_TO_HA.items()}
 
+SVC_SET_SYSTEM_MODE_SCHEMA = vol.Schema(
+    vol.Any(
+        cv.make_entity_service_schema(
+            {
+                vol.Required(ATTR_MODE): vol.In(
+                    [
+                        SystemMode.AUTO,
+                        SystemMode.HEAT_OFF,
+                        SystemMode.RESET,
+                    ]
+                )
+            }
+        ),
+        cv.make_entity_service_schema(
+            {
+                vol.Required(ATTR_MODE): vol.In([SystemMode.ECO_BOOST]),
+                vol.Optional(ATTR_DURATION, default=timedelta(hours=1)): vol.All(
+                    cv.time_period,
+                    vol.Range(min=timedelta(hours=1), max=timedelta(hours=24)),
+                ),
+            }
+        ),
+        cv.make_entity_service_schema(
+            {
+                vol.Required(ATTR_MODE): vol.In(
+                    [
+                        SystemMode.AWAY,
+                        SystemMode.CUSTOM,
+                        SystemMode.DAY_OFF,
+                        SystemMode.DAY_OFF_ECO,
+                    ]
+                ),
+                vol.Optional(ATTR_PERIOD, default=timedelta(days=0)): vol.All(
+                    cv.time_period,
+                    vol.Range(min=timedelta(days=0), max=timedelta(days=99)),
+                ),  # 0 means until the end of the day
+            }
+        ),
+    )
+)
+
+SVC_PUT_ZONE_TEMP_SCHEMA = cv.make_entity_service_schema(
+    {
+        vol.Required(ATTR_TEMPERATURE): vol.All(
+            vol.Coerce(float), vol.Range(min=-20, max=99)
+        ),
+    }
+)
+
+SVC_SET_ZONE_CONFIG_SCHEMA = cv.make_entity_service_schema(
+    {
+        vol.Optional(ATTR_MAX_TEMP, default=35): vol.All(
+            cv.positive_float, vol.Range(min=21, max=35)
+        ),
+        vol.Optional(ATTR_MIN_TEMP, default=5): vol.All(
+            cv.positive_float, vol.Range(min=5, max=21)
+        ),
+        vol.Optional(ATTR_LOCAL_OVERRIDE, default=True): cv.boolean,
+        vol.Optional(ATTR_OPENWINDOW, default=True): cv.boolean,
+        vol.Optional(ATTR_MULTIROOM, default=True): cv.boolean,
+    }
+)
+
+SVC_SET_ZONE_MODE_SCHEMA = vol.Schema(
+    vol.Any(
+        cv.make_entity_service_schema(
+            {
+                vol.Required(ATTR_MODE): vol.In([ZoneMode.SCHEDULE]),
+            }
+        ),
+        cv.make_entity_service_schema(
+            {
+                vol.Required(ATTR_MODE): vol.In(
+                    [ZoneMode.PERMANENT, ZoneMode.ADVANCED]
+                ),
+                vol.Optional(ATTR_SETPOINT, default=21): vol.All(
+                    cv.positive_float, vol.Range(min=5, max=30)
+                ),
+            }
+        ),
+        cv.make_entity_service_schema(
+            {
+                vol.Required(ATTR_MODE): vol.In([ZoneMode.TEMPORARY]),
+                vol.Optional(ATTR_SETPOINT, default=21): vol.All(
+                    cv.positive_float, vol.Range(min=5, max=30)
+                ),
+                vol.Exclusive(ATTR_UNTIL, ATTR_UNTIL): cv.datetime,
+                vol.Exclusive(ATTR_DURATION, ATTR_UNTIL): vol.All(
+                    cv.time_period,
+                    vol.Range(min=timedelta(minutes=5), max=timedelta(days=1)),
+                ),
+            }
+        ),
+    )
+)
+
+SVC_SET_ZONE_SCHEDULE_SCHEMA = cv.make_entity_service_schema(
+    {
+        vol.Required(ATTR_SCHEDULE): cv.string,
+    }
+)
+
 
 async def async_setup_platform(
     hass: HomeAssistant,
@@ -104,39 +227,56 @@ async def async_setup_platform(
 ) -> None:
     """Create climate entities for CH/DHW (heat) & HVAC."""
 
-    def entity_factory(entity_class, broker, device):  # TODO: deprecate
-        return entity_class(broker, device)
+    broker: RamsesBroker = hass.data[DOMAIN][BROKER]
+
+    if not broker._services.get(PLATFORM):
+        broker._services[PLATFORM] = True
+
+        platform = entity_platform.async_get_current_platform()
+
+        platform.async_register_entity_service(
+            SVC_RESET_SYSTEM_MODE, {}, "svc_reset_system_mode"
+        )
+        platform.async_register_entity_service(
+            SVC_SET_SYSTEM_MODE, SVC_SET_SYSTEM_MODE_SCHEMA, "svc_set_system_mode"
+        )
+        platform.async_register_entity_service(
+            SVC_PUT_ZONE_TEMP, SVC_PUT_ZONE_TEMP_SCHEMA, "put_zone_temp"
+        )
+        platform.async_register_entity_service(
+            SVC_SET_ZONE_CONFIG, SVC_SET_ZONE_CONFIG_SCHEMA, "svc_set_zone_config"
+        )
+        platform.async_register_entity_service(
+            SVC_RESET_ZONE_CONFIG, {}, "svc_reset_zone_config"
+        )
+        platform.async_register_entity_service(
+            SVC_SET_ZONE_MODE, SVC_SET_ZONE_MODE_SCHEMA, "svc_set_zone_mode"
+        )
+        platform.async_register_entity_service(
+            SVC_RESET_ZONE_MODE, {}, "svc_reset_zone_mode"
+        )
+        platform.async_register_entity_service(
+            SVC_GET_ZONE_SCHEDULE, {}, "svc_get_zone_schedule"
+        )
+        platform.async_register_entity_service(
+            SVC_SET_ZONE_SCHEDULE, SVC_SET_ZONE_SCHEDULE_SCHEMA, "svc_set_zone_schedule"
+        )
 
     if discovery_info is None:
         return
 
-    platform = entity_platform.async_get_current_platform()
-
-    broker = hass.data[DOMAIN][BROKER]
     new_entities = []
 
     if discovery_info.get("fans"):
-        if not broker._services.get(f"{PLATFORM}_hvac"):
-            broker._services[f"{PLATFORM}_hvac"] = True
-            # for name, schema in SVCS_CLIMATE_HVAC.items():
-            #     platform.async_register_entity_service(name, schema, f"svc_{name}")
-
         for fan in discovery_info["fans"]:
             new_entities.append(RamsesHvac(broker, fan))
 
     if discovery_info.get("ctls") or discovery_info.get("zons"):
-        if not broker._services.get(f"{PLATFORM}_heat"):
-            broker._services[f"{PLATFORM}_heat"] = True
-            for name, schema in SVCS_CLIMATE_EVO_TCS.items():
-                platform.async_register_entity_service(name, schema, f"svc_{name}")
-            for name, schema in SVCS_CLIMATE_EVO_ZONE.items():
-                platform.async_register_entity_service(name, schema, f"svc_{name}")
-
         for tcs in discovery_info.get("ctls", []):
-            new_entities.append(entity_factory(RamsesController, broker, tcs))
+            new_entities.append(RamsesController(broker, tcs))
 
         for zone in discovery_info.get("zons", []):
-            new_entities.append(entity_factory(RamsesZone, broker, zone))
+            new_entities.append(RamsesZone(broker, zone))
 
     if new_entities:
         async_add_entities(new_entities)
@@ -367,7 +507,7 @@ class RamsesZone(RamsesEntity, ClimateEntity):
 
         if self._device.tcs.system_mode is None:
             return None  # unable to determine
-        # if self._device.tcs.system_mode[CONF_SYSTEM_MODE] in MODE_TCS_TO_HA:
+        # if self._device.tcs.system_mode[SZ_SYSTEM_MODE] in MODE_TCS_TO_HA:
         if self._device.tcs.system_mode[SZ_SYSTEM_MODE] in (
             SystemMode.AWAY,
             SystemMode.HEAT_OFF,
