+++ custom_components/ramses_cc/const.py
@@ -28,6 +28,7 @@ ATTR_BATTERY_LEVEL = "battery_level"
 ATTR_CO2_LEVEL = "co2_level"
 ATTR_COMMAND = "command"
 ATTR_DELAY_SECS = "delay_secs"
+ATTR_DEVICE_ID = "device_id"
 ATTR_DIFFERENTIAL = "differential"
 ATTR_DURATION = "duration"
 ATTR_PERIOD = "period"
